+++ x/wasm/internal/types/types.go
@@ -70,6 +70,7 @@ type ContractCodeHistoryOperationType string
 const (
 	InitContractCodeHistoryType    ContractCodeHistoryOperationType = "Init"
 	MigrateContractCodeHistoryType ContractCodeHistoryOperationType = "Migrate"
+	GenesisContractCodeHistoryType ContractCodeHistoryOperationType = "Genesis"
 )
 
 var AllCodeHistoryTypes = []ContractCodeHistoryOperationType{InitContractCodeHistoryType, MigrateContractCodeHistoryType}
@@ -90,7 +91,7 @@ type ContractInfo struct {
 	// never show this in query results, just use for sorting
 	// (Note: when using json tag "-" amino refused to serialize it...)
 	Created             *AbsoluteTxPosition        `json:"created,omitempty"`
-	ContractCodeHistory []ContractCodeHistoryEntry `json:"contract_code_history"`
+	ContractCodeHistory []ContractCodeHistoryEntry `json:"contract_code_history,omitempty"`
 }
 
 func (c *ContractInfo) AddMigration(ctx sdk.Context, codeID uint64, msg []byte) {
@@ -122,14 +123,18 @@ func (c *ContractInfo) ValidateBasic() error {
 	if err := validateLabel(c.Label); err != nil {
 		return sdkerrors.Wrap(err, "label")
 	}
-	if c.Created == nil {
-		return sdkerrors.Wrap(ErrEmpty, "created")
-	}
-	if err := c.Created.ValidateBasic(); err != nil {
-		return sdkerrors.Wrap(err, "created")
-	}
-	// TODO: validate history
 	return nil
+}
+
+// ResetFromGenesis resets contracts timestamp and history.
+func (c *ContractInfo) ResetFromGenesis(ctx sdk.Context) {
+	c.Created = NewAbsoluteTxPosition(ctx)
+	h := ContractCodeHistoryEntry{
+		Operation: GenesisContractCodeHistoryType,
+		CodeID:    c.CodeID,
+		Updated:   c.Created,
+	}
+	c.ContractCodeHistory = []ContractCodeHistoryEntry{h}
 }
 
 // AbsoluteTxPosition can be used to sort contracts